// frontend/src/app/shared/models/diagnosis.model.ts
export class Diagnosis {

    address_patient: string;
    address_doctor: string;
    name: string;
    last_name: string;
    last_name_2: string;
    comorb: string;
    age: Int16Array;
    weigth: Int16Array;
    description: string;        // diagnostic
    observations: string;
    files: string[];
  }
  